mud file for gpe contributed by Graham Cobb  * libmimedir-gnome  * libgpewidget  * libgpepimc  * libeventdb  * libtododb  * libgpevtype  * libcontactsdb  * gpesyncd  * gpe-icons  * libschedule  * libxsettings-client  * libhandoff  * libsoup  * gpe-calendar  * gpe-contacts  * gpe-todo
As there is no recent "release" for GPE (nor one planned), these are set up to build from the SVN using a
particular date.  The date is then reflected in the version of the generated package, to allow easy tracking by the
users (which is why I use dates rather than revisions).  The date in the packages as submitted should be a reasonable
choice for an initial build. Although there are several bugs outstanding, there seems no particular point in delaying.
If/when some of the bugs are fixed we can update to a newer date.